Fixed a number of double and long castings
git-svn-id: https://satscm.esac.esa.int/esasky/esasky-app/trunk@5635 aa7853a0-da02-4334-bf23-07ffea9f4f01

+++ esasky-cl/src/main/java/esac/archive/esasky/cl/web/client/status/GUISessionStatus.java
@@ -224,7 +224,7 @@ public class GUISessionStatus {
 			}
 			UrlUtils.updateURLWithoutReloadingJS(UrlUtils.getUrlForCurrentState());
 			Date expires = new Date();
-			long milliseconds = 120 * 24 * 60 * 60 * 1000;
+			long milliseconds = ((long) 120)  * 24 * 60 * 60 * 1000;
 			expires.setTime(expires.getTime() + milliseconds);
 			Cookies.setCookie(EsaSkyWebConstants.SCI_MODE_COOKIE, Boolean.toString(isInScienceMode), expires);
 		}
